// src/game/TimePilot.types.ts
export type Heading = number;

export interface Coordinates {
  posX: number;
  posY: number;
}
// ...
export interface BulletData extends Coordinates {
  heading: Heading;
  size: number;
  velocity: number;
  color: string;
}
// ...
export interface BulletInstance {
  removeMe: boolean;
  getData: (key?: keyof BulletData) => BulletData | BulletData[keyof BulletData] | undefined;
  setData: (key: keyof BulletData, value: BulletData[keyof BulletData]) => boolean;
  setLevel: (level: number) => boolean;
  reposition: () => void;
  render: () => void;
}

export interface BulletFactoryInstance {
  create: (
    originX: number,
    originY: number,
    heading: Heading,
    size: number,
    velocity: number,
    color: string
  ) => void;
  getCount: () => number;
  getData: () => BulletData[];
  cleanup: () => void;
  reposition: () => void;
  render: () => void;
  clearAll: () => void;
}